farming-bot/coinbot/requirements.txt
undetected-chromedriver
webdriver-manager
selenium==4.15.2
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
fake-useragent==1.4.0
pytesseract==0.3.10
Pillow==10.1.0
psycopg2-binary
flask==3.0.0
